楽観ロック処理修正

+++ src/main/java/com/example/chat/repository/impl/CommentLikeRepositoryImpl.java
@@ -1,10 +1,13 @@
 package com.example.chat.repository.impl;
 
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Repository;
 
+import com.example.chat.exception.ConflictException;
 import com.example.chat.model.CommentLike;
 import com.example.chat.repository.CommentLikeRepository;
 import com.example.chat.repository.mapper.CommentLikeMapper;
+import com.example.chat.util.MessageCode;
 
 /**
  * コメントいいねリポジトリ
@@ -17,7 +20,7 @@ public class CommentLikeRepositoryImpl implements CommentLikeRepository {
 
 	/** コメントいいねマッパー */
 	private final CommentLikeMapper commentLikeMapper;
-	
+
 	/**
 	 * コンストラクタ
 	 * 
@@ -29,11 +32,20 @@ public class CommentLikeRepositoryImpl implements CommentLikeRepository {
 
 	@Override
 	public int insert(CommentLike commentLike) {
-		return this.commentLikeMapper.insert(commentLike);
+		try {
+			return this.commentLikeMapper.insert(commentLike);
+		} catch (DuplicateKeyException e) {
+			throw new ConflictException(MessageCode.CONFLICT_INSERT_LIKE, e);
+		}
+
 	}
 
 	@Override
 	public int delete(CommentLike commentLike) {
+		int resultCount = this.commentLikeMapper.delete(commentLike);
+		if (resultCount == 0) {
+			throw new ConflictException(MessageCode.CONFLICT_DELETE_LIKE);
+		}
 		return this.commentLikeMapper.delete(commentLike);
 	}
